requirements.txt
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
EasyProcess==1.1
entrypoint2==1.1
et_xmlfile==2.0.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
idna==3.10
loguru==0.7.3
msoffcrypto-tool==5.4.2
numpy==2.3.1
olefile==0.47
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pandas-stubs==2.3.0.250703
patool==4.0.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyunpack==0.3
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.41
tqdm==4.67.1
types-pytz==2025.2.0.20250516
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
win32_setctime==1.2.0
xlrd==2.0.2
